Move selection hooks to list directory
