+++ packages/ra-core/src/controller/index.ts
@@ -1,16 +1,13 @@
-import useExpanded from './useExpanded';
 import useFilterState from './useFilterState';
 import useSortState, { SortProps } from './useSortState';
 import usePaginationState, { PaginationHookResult } from './usePaginationState';
 
 import { useCheckMinimumRequiredProps } from './checkMinimumRequiredProps';
-import { useUnselectAll } from './useUnselectAll';
 
 export type { PaginationHookResult, SortProps };
 
 export {
     useCheckMinimumRequiredProps,
-    useExpanded,
     useFilterState,
     usePaginationState,
     useSortState,
@@ -26,6 +23,3 @@ export * from './record';
 export * from './saveContext';
 export * from './show';
 export * from './useReference';
-export * from './useRecordSelection';
-export * from './useUnselect';
-export * from './useUnselectAll';
