+++ packages/ra-core/src/controller/list/index.ts
@@ -13,3 +13,7 @@ export * from './useListPaginationContext';
 export * from './useListParams';
 export * from './useListSortContext';
 export * from './queryReducer';
+export * from './useRecordSelection';
+export * from './useUnselect';
+export * from './useUnselectAll';
+export * from './useExpanded';
